correcoes e firebase
