+++ package.json
@@ -9,6 +9,7 @@
     "@testing-library/user-event": "^7.1.2",
     "enzyme": "^3.11.0",
     "enzyme-adapter-react-16.1": "^1.6.2",
+    "firebase": "^7.14.1",
     "formik": "^2.1.4",
     "jest-enzyme": "^7.1.2",
     "moment": "^2.24.0",
